+++ src/main/java/racing/view/Input.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.view;
 
 import java.util.Scanner;
