+++ src/test/java/racing/GameTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racing.domain.Car;
+import racing.domain.Game;
 
 import java.util.Arrays;
 import java.util.List;
